refactor: moved indexedDB model to separate file and local storage same
